Fixes #4815: Divide RuleEditForm into three files (editform, compliance, compliance popup)
