+++ rudder-web/src/main/webapp/templates-hidden/components/ComponentRuleEditForm.html
@@ -90,22 +90,26 @@
 	<div class="groupedEditZone">
 	  <fieldset>
     <legend>Rule information</legend>
-		<div id="nameField">
-		Here comes the name field
-		</div>
-           <br />
-		<div id="categoryField">
-		Here comes the category field
-		</div>
-           <br />
-		<div id="shortDescriptionField">
-		Here comes the shortDescription field
-		</div>
-		   <br/>
-		<div id="longDescriptionField">
-		Here comes the longDescription field
-		</div>
-	    <br />
+        <div>
+          <b>Name: </b>
+          <span id="ruleName"/>
+        </div>
+        <br />
+        <div>
+          <b>Category: </b>
+          <span id="ruleCategory"/>
+        </div>
+        <br />
+        <div>
+          <b >Short description: </b>
+          <span id="ruleShortDescription"/>
+        </div>
+        <br />
+        <div >
+          <b>Long description: </b>
+          <span id="ruleLongDescription"/>
+        </div>
+        <br />
         <div class="RudderIdField">
             <b>Rudder ID:</b>
             <span id="rudderID">
